TASK: Create Soundfont repository
This is a collection of SF2 I've picked
up across the internet looking for some
suitable to use with Teensy Audio WaveTable

Curiously the best "small one" with reasonable
sounding instruments that should fit in T4.0
is actually "gm.sf2" which is the one that comes
with Windows. So will decode to source files
in decode\gm directory over time.